pyproject.toml
[build-system]
requires = ["poetry>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "markdown"
version = "0.1.0"
description = "Ein Beispielprojekt"
authors = ["Falk Schettler <falk.schettler@gmx.de>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
flake8 = "^7.1.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-cov = "^6.0"
coveralls = "^4.0.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"

